+++ CodeHub.Core/Services/IServiceConstructor.cs
@@ -6,5 +6,13 @@ namespace CodeHub.Core.Services
     {
         object Construct(Type type);
     }
+
+    public static class ServiceConstructorExtensions
+    {
+        public static T Construct<T>(this IServiceConstructor serviceConstructor)
+        {
+            return (T)serviceConstructor.Construct(typeof(T));
+        }
+    }
 }
