--- CodeHub.Core/Services/ServiceConstructorExtensions.cs
@@ -1,11 +1,0 @@
-﻿namespace CodeHub.Core.Services
-{
-    public static class ServiceConstructorExtensions
-    {
-        public static T Construct<T>(this IServiceConstructor serviceConstructor)
-        {
-            return (T)serviceConstructor.Construct(typeof(T));
-        }
-    }
-}
-
